# rope-manager-backend/openapi.yaml
openapi: 3.0.3
info:
  title: 绳包管理器后端 API
  description: |
    绳包管理器后端服务的完整API文档
    
    ## 认证方式
    系统支持两种认证方式，按优先级顺序：
    1. **Authorization 头部认证（推荐）**: `Authorization: Bearer <token>`
    2. **Cookie 认证**: `Cookie: auth_token=<token>`
    
    ## 权限等级
    - **普通用户 (user)**: 基础功能使用权限
    - **元老 (elder)**: 拥有部分管理权限，可审核资源
    - **版主 (moderator)**: 拥有内容管理权限
    - **管理员 (admin)**: 拥有所有系统权限
    
    ## API 响应格式
    所有API统一返回格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
    - code: 状态码，0表示成功，其他表示错误
    - message: 响应消息
    - data: 响应数据（可选）
    
  version: 1.0.0
  contact:
    name: 绳包管理器API支持
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:15201/api/v1
    description: 开发服务器 - API端点
  - url: http://127.0.0.1:15201
    description: 开发服务器 - 健康检查端点
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
        message:
          type: string
          description: 响应消息
        data:
          description: 响应数据（可选）
      required:
        - code
        - message

    # 用户相关模型
...
    CreatePackageRequest:
      type: object
      properties:
        name:
          type: string
        author:
          type: string
        version:
          type: string
        description:
          type: string
        category_id:
          type: integer
        file_url:
          type: string
        tags:
          type: array
          items:
            type: string
        is_pinned:
          type: boolean
        is_featured:
          type: boolean
      required:
        - name
        - author
...
    CreateResourceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        file_url:
          type: string
        cover_url:
          type: string
      required:
        - title
        - file_url
    PackageListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 评论相关模型
...
    CommentListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 帖子相关模型
...
paths:
...
  /packages:
    get:
      tags:
        - 绳包管理
      summary: 获取绳包列表
      description: 获取绳包列表，支持分页和筛选
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, Active, Rejected, Inactive, Deleted]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PackageListResponse'
  /packages/pending:
    get:
      tags:
        - 绳包管理
      summary: 获取待审核资源
      description: 获取待审核的资源列表（管理员和元老权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PackageListResponse'
        '403':
          description: 权限不足
...
  /packages/user-submit:
    post:
      tags:
        - 绳包管理
      summary: 用户提交资源
      description: 普通用户提交资源，自动设置为待审核状态
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          package_id:
                            type: integer
        '401':
          description: 未授权
  /packages/admin-create:
    post:
      tags:
        - 绳包管理
      summary: 管理员创建资源
      description: 管理员或元老创建资源，可设置任意状态
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePackageRequest'
      responses:
        '200':
          description: 创建成功
        '403':
          description: 权限不足
  /packages/{id}:
    get:
      tags:
        - 绳包管理
      summary: 获取绳包详情
      description: 根据ID获取绳包详细信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Package'
        '404':
          description: 绳包不存在

    put:
      tags:
        - 绳包管理
      summary: 更新绳包信息
      description: 更新绳包信息（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePackageRequest'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 权限不足
        '404':
          description: 绳包不存在

    delete:
      tags:
        - 绳包管理
      summary: 删除绳包
      description: 删除绳包（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
        '404':
          description: 绳包不存在
  /packages/{id}/download:
    get:
      tags:
        - 绳包管理
      summary: 下载绳包
      description: 下载绳包文件
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 开始下载
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: 文件不存在
        '429':
          description: 下载频率限制
  /packages/{id}/upload:
    post:
      tags:
        - 绳包管理
      summary: 上传绳包文件
      description: 上传绳包文件（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: 上传成功
        '403':
          description: 权限不足
        '404':
          description: 绳包不存在
  /packages/{id}/comments:
    get:
      tags:
        - 绳包管理
      summary: 获取绳包评论
      description: 获取指定绳包的评论列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'
  /packages/{id}/review:
    post:
      tags:
        - 绳包管理
      summary: 审核资源
      description: 审核资源（管理员和元老权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewResourceRequest'
      responses:
        '200':
          description: 审核成功
        '400':
          description: 审核失败
        '403':
          description: 权限不足
        '404':
          description: 资源不存在

  # 评论管理接口
...
  /categories:
    get:
      tags:
        - 分类管理
      summary: 获取分类列表
      description: 获取所有分类列表，包含每个分类的资源数量
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: integer
                                name:
                                  type: string
                                description:
                                  type: string
                                enabled:
                                  type: boolean
                                subscription_locked:
                                  type: boolean
                                created_at:
                                  type: string
                                  format: date-time
                                updated_at:
                                  type: string
                                  format: date-time
                                count:
                                  type: integer
        '500':
          description: 服务器错误

    post:
      tags:
        - 分类管理
      summary: 创建分类
      description: 创建新的分类（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                enabled:
                  type: boolean
                  default: true
                subscription_locked:
                  type: boolean
                  default: false
              required:
                - name
      responses:
        '200':
          description: 创建成功
        '400':
          description: 创建失败
  /categories/{id}:
    get:
      tags:
        - 分类管理
      summary: 获取分类详情
      description: 根据ID获取分类详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
        '404':
          description: 分类不存在

    put:
      tags:
        - 分类管理
      summary: 更新分类
      description: 更新分类信息（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                enabled:
                  type: boolean
                subscription_locked:
                  type: boolean
      responses:
        '200':
          description: 更新成功
        '400':
          description: 更新失败

    delete:
      tags:
        - 分类管理
      summary: 删除分类
      description: 删除分类（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '500':
          description: 删除失败

  # 社区动态接口
...